Introduce an optional external_uri on the instance.
It is used to define a hostname, path and protocol that is exposed
to the end user in contrast to the URL used by the local webserver.

By default, this is the same as the base_uri. Setting something
different here only makes sense if you have an additional proxy
component in front of your webserver.

For example, you could set this if you have an SSL accelerator in front of your Webserver:

"base_uri": "http://example.com",
"external_uri": "https://example.com"

The instance will now only have a single virtual host listening on
port 80 using simple HTTP. Depending on where your proxy / acceler-
ator lives, it might be necessary to perform the redirect to HTTPS on your proxy as requests from localhost will not be forwarded by
us to allow easier access for e.g. cron scripts.

In conclusion: the base_uri defines what will be setup as a vhost
on your webserver, the external_uri defines what the user sees. If
they differ, you have to make sure the necessary rewrites are
properly performed by your additional infrastructure.